chore(audit): add auditor base
