+++ kayle_innate/Cargo.toml
@@ -30,7 +30,7 @@ js-sys = "0.3.64"
 selectors = "=0.21.0"
 smallvec = "0.6"
 ego-tree = "0.6.2"
-victor_tree = { version = "0.0.5" }
+victor_tree = { version = "0.0.7" }
 markup5ever = "0.11.0"
 
 [dependencies.cssparser]
