+++ kayle_innate/src/engine/audit/mod.rs
@@ -1,2 +1,6 @@
+/// the auditor
+pub mod auditor;
+/// the node tree
+pub mod tree;
 /// WCAG audit
 pub mod wcag;
