+++ kayle_innate/src/engine/styles/errors.rs
@@ -1,12 +1,12 @@
-use cssparser::{CowRcStr, ParseError};
+// use cssparser::{CowRcStr, ParseError};
 use selectors::parser::SelectorParseErrorKind;
 
-pub type PropertyParseError<'i> = ParseError<'i, PropertyParseErrorKind<'i>>;
+// pub type PropertyParseError<'i> = ParseError<'i, PropertyParseErrorKind<'i>>;
 
-pub enum PropertyParseErrorKind<'i> {
-    UnknownProperty(CowRcStr<'i>),
-    UnknownUnit(CowRcStr<'i>),
-}
+// pub enum PropertyParseErrorKind<'i> {
+//     UnknownProperty(CowRcStr<'i>),
+//     UnknownUnit(CowRcStr<'i>),
+// }
 
 pub enum RuleParseErrorKind<'i> {
     Selector(SelectorParseErrorKind<'i>),
