+++ kayle_innate/src/lib.rs
@@ -1,22 +1,24 @@
 #[cfg(feature = "wee_alloc")]
 #[global_allocator]
 static ALLOC: wee_alloc::WeeAlloc = wee_alloc::WeeAlloc::INIT;
-use wasm_bindgen::prelude::*;
+pub use wasm_bindgen::prelude::*;
 
 #[macro_use]
 extern crate lazy_static;
 
+/// the main engine for audits
 mod engine;
+/// locales
 mod i18n;
+/// used for auto-kayle to gather all links in a page.
+mod links;
+/// helpers
 mod utils;
 
-use case_insensitive_string::CaseInsensitiveString;
+#[cfg(feature = "accessibility")]
+use crate::engine::audit::auditor::Auditor;
 #[cfg(feature = "accessibility")]
 use scraper_forky::ElementRef;
-#[cfg(feature = "accessibility")]
-use std::collections::BTreeMap;
-use std::collections::HashSet;
-use utils::{convert_abs_path, convert_base_path, domain_name, set_panic_hook};
 
 #[cfg(feature = "accessibility")]
 #[wasm_bindgen]
@@ -38,263 +40,35 @@ macro_rules! console_log {
 }
 
 #[wasm_bindgen]
-/// setup a structure tree alg for parsing and find links in document. Allow user to perform hybrid audits realtime.
-pub fn get_document_links(res: &str, domain: &str) -> Box<[JsValue]> {
-    set_panic_hook();
-
-    lazy_static! {
-        /// include only list of resources
-        static ref ONLY_RESOURCES: HashSet<CaseInsensitiveString> = {
-            let mut m: HashSet<CaseInsensitiveString> = HashSet::with_capacity(14);
-
-            m.extend([
-                "html", "htm", "asp", "aspx", "php", "jps", "jpsx",
-                // handle .. prefix for urls ending with an extra ending
-                ".html", ".htm", ".asp", ".aspx", ".php", ".jps", ".jpsx",
-            ].map(|s| s.into()));
-
-            m
-        };
-    }
-
-    let links = match url::Url::parse(domain) {
-        Ok(base) => {
-            let base_url = convert_base_path(base);
-            let base_domain = domain_name(&base_url);
-            let parent_host_scheme = base_url.scheme();
-            let parent_host = base_url.host_str().unwrap_or_default();
-
-            let h = scraper_forky::Html::parse_fragment(res);
-
-            h.tree
-                .into_iter()
-                .filter_map(|node| {
-                    if let Some(element) = node.as_element() {
-                        if element.name() == "a" {
-                            match element.attr("href") {
-                                Some(link) => {
-                                    let mut abs = convert_abs_path(&base_url, link);
-                                    let mut can_process = match abs.host_str() {
-                                        Some(host) => parent_host.ends_with(host),
-                                        _ => false,
-                                    };
-
-                                    let process = if can_process {
-                                        if abs.scheme() != parent_host_scheme {
-                                            let _ = abs.set_scheme(parent_host_scheme);
-                                        }
-
-                                        let hchars = abs.path();
-
-                                        if let Some(position) = hchars.find('.') {
-                                            let resource_ext = &hchars[position + 1..hchars.len()];
-
-                                            if !ONLY_RESOURCES.contains::<CaseInsensitiveString>(
-                                                &resource_ext.into(),
-                                            ) {
-                                                can_process = false;
-                                            }
-                                        }
-
-                                        if can_process
-                                            && (base_domain.is_empty()
-                                                || base_domain == domain_name(&abs))
-                                        {
-                                            Some(JsValue::from_str(&abs.as_str()))
-                                        } else {
-                                            None
-                                        }
-                                    } else {
-                                        None
-                                    };
-
-                                    process
-                                }
-                                _ => None,
-                            }
-                        } else {
-                            None
-                        }
-                    } else {
-                        None
-                    }
-                })
-                .collect::<Vec<_>>()
-        }
-        _ => {
-            let h = scraper_forky::Html::parse_fragment(res);
-
-            h.tree
-                .into_iter()
-                .filter_map(|node| {
-                    if let Some(element) = node.as_element() {
-                        if element.name() == "a" {
-                            match element.attr("href") {
-                                Some(link) => {
-                                    // TODO: validate only web links
-                                    Some(JsValue::from_str(&link))
-                                }
-                                _ => None,
-                            }
-                        } else {
-                            None
-                        }
-                    } else {
-                        None
-                    }
-                })
-                .collect::<Vec<_>>()
-        }
-    };
-
-    links.into_boxed_slice()
-}
-
-// RUST_LOG=info wasm-pack test --firefox --headless --features accessibility --release
-#[cfg(feature = "accessibility")]
-/// try to fix all possible issues using a spec against the tree.
-pub fn parse_accessibility_tree(
-    html: &scraper_forky::Html,
-    // todo: return the nodes with a tuple of the layout node and the element node
-) -> std::collections::BTreeMap<&str, Vec<ElementRef<'_>>> {
-    // use taffy::prelude::*;
-    // // todo: use optional variable for clips or layout creation
-    // let mut taffy = Taffy::new();
-
-    // let header_node = taffy
-    //     .new_leaf(Style {
-    //         size: Size {
-    //             width: points(800.0),
-    //             height: points(100.0),
-    //         },
-    //         ..Default::default()
-    //     })
-    //     .unwrap();
-
-    // let body_node = taffy
-    //     .new_leaf(Style {
-    //         size: Size {
-    //             width: points(800.0),
-    //             height: auto(),
-    //         },
-    //         flex_grow: 1.0,
-    //         ..Default::default()
-    //     })
-    //     .unwrap();
-
-    // let root_node = taffy
-    //     .new_with_children(
-    //         Style {
-    //             flex_direction: FlexDirection::Column,
-    //             size: Size {
-    //                 width: points(800.0),
-    //                 height: points(600.0),
-    //             },
-    //             ..Default::default()
-    //         },
-    //         &[header_node, body_node],
-    //     )
-    //     .unwrap();
-
-    // // Call compute_layout on the root of your tree to run the layout algorithm
-    // taffy.compute_layout(root_node, Size::MAX_CONTENT).unwrap();
-    // console_log!("Header Layout {:?}", taffy.layout(header_node).unwrap());
-    // We can get the x,y, and height, width of the element on proper tree insert
-
-    let t = now();
-    // parse doc will start from html downwards
-    // accessibility tree for ordered element mappings
-    let mut accessibility_tree: BTreeMap<&str, Vec<ElementRef<'_>>> =
-        BTreeMap::from([("title".into(), Default::default())]);
-
-    for node in html.tree.nodes() {
-        match scraper_forky::element_ref::ElementRef::wrap(node) {
-            Some(element) => {
-                accessibility_tree
-                    .entry(element.value().name())
-                    .and_modify(|n| n.push(element))
-                    .or_insert(Vec::from([element]));
-            }
-            _ => (),
-        };
-    }
-
-    console_log!("Scraper Parser: duration {:?}ms", now() - t);
-    // console_log!("Getting tree links {:?}", accessibility_tree.get("a"));
-    // console_log!("Tree {:?}", accessibility_tree);
-
-    accessibility_tree
-}
-
-#[wasm_bindgen]
 #[cfg(feature = "accessibility")]
 /// audit a web page passing the html and css rules.
-pub fn _audit_not_ready(html: &str, _css_rules: &str) -> Result<JsValue, JsValue> {
+pub fn _audit_not_ready(html: &str, css_rules: &str) -> Result<JsValue, JsValue> {
+    let t = now();
     let document = scraper_forky::Html::parse_document(html);
-    let _tree = parse_accessibility_tree(&document);
-    let _author = {
-        let mut author = victor_tree::style::StyleSetBuilder::new();
-        if !_css_rules.is_empty() {
-            author.add_stylesheet(_css_rules);
-        } else {
-            use markup5ever::local_name;
-            match _tree.get("style") {
-                Some(styles) => {
-                    for node in styles {
-                        // https://html.spec.whatwg.org/multipage/semantics.html#update-a-style-block
-                        if let Some(type_attr) = node.attr(&local_name!("type")) {
-                            if !type_attr.eq_ignore_ascii_case("text/css") {
-                                continue;
-                            }
-                            author.add_stylesheet(&node.inner_html())
-                        }
-                    }
-                }
-                _ => (),
-            }
-        }
-        author.finish()
-    };
-    let _audit = engine::audit::wcag::WCAG3AA::audit(&_tree);
-
-    // // TODO: build struct that can keep lifetimes
+    console_log!("Parse Document Time {:?}", now() - t);
     let mut nth_index_cache = selectors::NthIndexCache::from(Default::default());
-    let mut match_context = selectors::matching::MatchingContext::new(
-        selectors::matching::MatchingMode::Normal,
-        None,
-        Some(&mut nth_index_cache),
-        selectors::matching::QuirksMode::NoQuirks,
-        // selectors::matching::NeedsSelectorFlags::No,
-        // selectors::matching::IgnoreNthChildForInvalidation::No,
+    let auditor = Auditor::new(
+        &document,
+        &css_rules,
+        engine::styles::css_cache::build_matching_context(&mut nth_index_cache),
     );
+    let ttt = now();
+    let _audit = engine::audit::wcag::WCAG3AA::audit(&auditor);
+    console_log!("Audit Time {:?}", now() - ttt);
+    
+    // let mut _match_context = auditor.match_context;
 
-    for item in _tree {
-        for node in item.1 {
-            let parent_styles = match node.parent() {
-                Some(n) => match scraper_forky::element_ref::ElementRef::wrap(n) {
-                    Some(element) => {
-                        let _parent_styles = crate::engine::styles::style_for_element(
-                            &_author,
-                            &document,
-                            element,
-                            None,
-                            &mut match_context,
-                        );
-                        Some(_parent_styles)
-                    }
-                    _ => None,
-                },
-                _ => None,
-            };
-            let _style = crate::engine::styles::style_for_element(
-                &_author,
-                &document,
-                node,
-                parent_styles.as_deref(),
-                &mut match_context,
-            );
-        }
-    }
+    // for item in auditor.tree {
+    //     for node in item.1 {
+    //         let _style = victor_tree::style::cascade::style_for_element_ref(
+    //             &node,
+    //             &auditor.author,
+    //             &document,
+    //             &mut _match_context,
+    //         );
+    //         console_log!("{:?}", _style.as_ref().box_size())
+    //     }
+    // }
 
     // todo: map to JsValues instead of serde
     Ok(serde_wasm_bindgen::to_value(&_audit)?)
